requirements.txt
django
djangorestframework
pyverse
pylint
pylint-django